Move loading of Non-sdk projects to a helper class
