+++ src/extension/NonSdkProjectHelper.cs
@@ -1,0 +1,83 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Xml;
+
+namespace NUnit.Engine.Services.ProjectLoaders
+{
+    /// <summary>
+    /// Load a non-C++ project in the MsBuild format introduced with VS2005
+    /// </summary>
+    public static class NonSdkProjectHelper
+    {
+        public static void LoadProject(VSProject project, XmlDocument doc)
+        {
+            XmlNamespaceManager namespaceManager = new XmlNamespaceManager(doc.NameTable);
+            namespaceManager.AddNamespace("msbuild", "http://schemas.microsoft.com/developer/msbuild/2003");
+
+            XmlNodeList nodes = doc.SelectNodes("/msbuild:Project/msbuild:PropertyGroup", namespaceManager);
+            if (nodes == null) return;
+
+            XmlElement assemblyNameElement = (XmlElement)doc.SelectSingleNode("/msbuild:Project/msbuild:PropertyGroup/msbuild:AssemblyName", namespaceManager);
+            string assemblyName = assemblyNameElement == null ? project.Name : assemblyNameElement.InnerText;
+
+            XmlElement outputTypeElement = (XmlElement)doc.SelectSingleNode("/msbuild:Project/msbuild:PropertyGroup/msbuild:OutputType", namespaceManager);
+            string outputType = outputTypeElement == null ? "Library" : outputTypeElement.InnerText;
+
+            if (outputType == "Exe" || outputType == "WinExe")
+                assemblyName = assemblyName + ".exe";
+            else
+                assemblyName = assemblyName + ".dll";
+
+            string commonOutputPath = null;
+            var explicitOutputPaths = new Dictionary<string, string>();
+
+            foreach (XmlElement configNode in nodes)
+            {
+                string name = GetConfigNameFromCondition(configNode);
+
+                XmlElement outputPathElement = (XmlElement)configNode.SelectSingleNode("msbuild:OutputPath", namespaceManager);
+                string outputPath = null;
+                if (outputPathElement != null)
+                    outputPath = outputPathElement.InnerText;
+
+                if (name == null)
+                {
+                    if (outputPathElement != null)
+                        commonOutputPath = outputPath;
+                    continue;
+                }
+
+                if (outputPathElement != null)
+                    explicitOutputPaths[name] = outputPath;
+
+                if (outputPath == null)
+                    outputPath = explicitOutputPaths.ContainsKey(name) ? explicitOutputPaths[name] : commonOutputPath;
+
+                if (outputPath != null && !project.ConfigNames.Contains(name))
+                    project.AddConfig(name, Path.Combine(outputPath.Replace("$(Configuration)", name), assemblyName));
+            }
+        }
+        private static string GetConfigNameFromCondition(XmlElement configNode)
+        {
+            string configurationName = null;
+            XmlAttribute conditionAttribute = configNode.Attributes["Condition"];
+            if (conditionAttribute != null)
+            {
+                string condition = conditionAttribute.Value;
+                if (condition.IndexOf("$(Configuration)") >= 0)
+                {
+                    int start = condition.IndexOf("==");
+                    if (start >= 0)
+                    {
+                        configurationName = condition.Substring(start + 2).Trim(new char[] { ' ', '\'' });
+                        int bar = configurationName.IndexOf('|');
+                        if (bar > 0)
+                            configurationName = configurationName.Substring(0, bar);
+                    }
+                }
+            }
+            return configurationName;
+        }
+    }
+}
